Implement AuthZ

+++ src/main/java/com/zuehlke/securesoftwaredevelopment/controller/PersonsController.java
@@ -1,6 +1,7 @@
 package com.zuehlke.securesoftwaredevelopment.controller;
 
 import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
+import com.zuehlke.securesoftwaredevelopment.config.SecurityUtil;
 import com.zuehlke.securesoftwaredevelopment.domain.Person;
 import com.zuehlke.securesoftwaredevelopment.domain.User;
 import com.zuehlke.securesoftwaredevelopment.repository.PersonRepository;
@@ -9,6 +10,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -34,6 +36,7 @@ public class PersonsController {
     }
 
     @GetMapping("/persons/{id}")
+    @PreAuthorize("hasAuthority('VIEW_PERSON')")
     public String person(@PathVariable int id, Model model, HttpSession httpSession) {
         model.addAttribute("CSRF_TOKEN", httpSession.getAttribute("CSRF_TOKEN"));
         model.addAttribute("person", personRepository.get("" + id));
@@ -41,6 +44,7 @@ public class PersonsController {
     }
 
     @GetMapping("/myprofile")
+    @PreAuthorize("hasAuthority('VIEW_MY_PROFILE')")
     public String self(Model model, Authentication authentication, HttpSession httpSession) {
         User user = (User) authentication.getPrincipal();
         model.addAttribute("CSRF_TOKEN", httpSession.getAttribute("CSRF_TOKEN"));
@@ -49,7 +53,13 @@ public class PersonsController {
     }
 
     @DeleteMapping("/persons/{id}")
-    public ResponseEntity<Void> person(@PathVariable int id) {
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
+    public ResponseEntity<Void> person(@PathVariable int id) throws AccessDeniedException {
+        User user = SecurityUtil.getCurrentUser();
+        if (user.getId() != id) {
+            throw new AccessDeniedException("Cannot delete another user");
+        }
+
         personRepository.delete(id);
         userRepository.delete(id);
 
@@ -57,6 +67,7 @@ public class PersonsController {
     }
 
     @PostMapping("/update-person")
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
     public String updatePerson(
             Person person,
             HttpSession httpSession,
@@ -66,17 +77,25 @@ public class PersonsController {
         if (!csrfToken.equals(expectedCsrfToken)) {
             throw new AccessDeniedException("Forbidden");
         }
+
+        User user = SecurityUtil.getCurrentUser();
+        if (user.getId() != Integer.parseInt(person.getId())) {
+            throw new AccessDeniedException("Cannot update another user");
+        }
+
         personRepository.update(person);
         return "redirect:/persons/" + person.getId();
     }
 
     @GetMapping("/persons")
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     public String persons(Model model) {
         model.addAttribute("persons", personRepository.getAll());
         return "persons";
     }
 
     @GetMapping(value = "/persons/search", produces = "application/json")
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     @ResponseBody
     public List<Person> searchPersons(@RequestParam String searchTerm) throws SQLException {
         return personRepository.search(searchTerm);
